astra-crud-extension/requirements.txt
Flask==3.0.1
gunicorn==21.2.0
astrapy==0.7.1
google-cloud-secret-manager==2.17.0
